requirements.txt
aiohttp==3.5.4
aredis==1.1.5
sanic_scheduler==1.0.7
sanic==18.12.0
async_timeout==3.0.1
ujson==1.35
PyMySQL==0.9.2
aelog==1.0.3
aiocache==0.11.1
aioredis==1.2.0
pymongo==3.9.0
asyncio_redis==0.15.1
motor==2.0.0
uvloop==0.12.2
aiomysql==0.0.20
aclients==1.0.0b40
beautifulsoup4==4.8.1
PyYAML==5.1.2
gunicorn
honcho
